+++ ChangeLog.md
@@ -1,2 +1,5 @@
+## Changes between o3d3xx-h5 0.1.0 and 0.1.1
+
+* Aware of the change from `xyzi_image` to `xyz_image` in o3d3xx-ros
 
 ## The initial release of o3d3xx-h5 was 0.1.0